Session Name, and Reload Bind
By using a session name, we ensure that we always attach to the default
Koding session. As a bonus, it means when we reload our config *(with
the newly added `C-a C-r` bind)*, we won't create a new session.
Currently it still tries to make a new session, but at least it fails..
which is better.. i'd still like it to check before it creates though.