Added Alias to Profile
Why? Well, when you SSH in profile is called but not bashrc. So,
i want to easily be able to launch tmux, and thus i added the
alias.

Why not auto launch? Mostly for KDApps. When you make a
`vmController.run()` call it runs the code on Koding and some kdapps
require/expect bash. Understandable, but we have to keep that in mind.
So, bash_profile can't always run tmux. Only bashrc can.