Powerline Quirk, UTF8, and Fish Bin
The powerline call was moved to the end of the file due to an inability
to make it work properly. See the code comment for further details.

UTF-8 is now enabled, for good measure (patched fonts)

The fish bin is different for my mac, so until i figure exactly how i
want to handle the different system locations/etc, i am just hardcoding
the location.