Fixed docker-shared/.ssh linking, and removed ssh/config usage.
The ssh config usage, as the given comment explains, is not possible
during the build process. We were trying to add a file to a directory
that just does not exist (ie, a shared directory that hasn't yet been
shared).

So, this will either need to be added post-run, or simply manually. For
now, i'm just adding them manually myself.