Fixed fish autocomplete, added full go install
Fish Autocomplete
The autocomplete for certain things (not sure on the difference) was
causing it to error out. Apparently `man` is required for this.. So, it
was added.

Go Install
Previously i was just adding the go bins manually. However, GOROOT then
becomes a non-existent location.. which causes some problems. So, i
added the traditional go install (fully), and then added the proper path
to it.