vim: fixed broken autocmd
For some unknown reason the autocmd was not working properly from within
the ftplugin. After exhaustive trial and error, an `autocmd! <cmd>` by
itself worked, but not the `augroup`. Moreover, the `augroup` worked
fine if called from init.vim instead of the ftplugin/rust.vim.

So, in the end the autocmd wasn't actually needed anyway, since the
ftplugin takes care of binding those per filetype anyway. So i just
removed the autocmds.