vim: added vim-go back and changed goto definition UX
using go for a project, so i'm adding my vim tools back.

Also tweaked the previously `<Leader>gd` to be `<Leader>gg`. The idea
being, a leader key has a main function (g for go to def, and k for
documentation), so gg goes to that definition, and kk goes to the
documentation.