nix(home/term): update unstable-pkgs
Note that it updated to `master`, as that seems to be the new thing
for unstable.

Also had to update following inputs, as they were failing with
nixUnstable messages. Which is what gave me the idea to use master.

Not sure if this is the right way to use master with nix, but it worked
/shrug.