nix: introduce `dev go` command
To easily hop in a fixed go development for repos
i don't own.

I'm not happy with the embedded lockfile, i'll probably change that in
the future. I went through quite a few hoops trying to find a way to
both link it to the root lockfile without actually referencing it
directly.

Finally i ended up linking to this repo because `nix develop` can't
resolve the double links that home manager is setting up..
so... yea, /shrug.